HOME-38 - Implement multiple agents
* Add dynamic menu rendering (when agent is added from agents config,
it should be visible on the menu)
