+++ controllers/agent.go
@@ -2,7 +2,7 @@ package controllers
 
 import (
     "net/http"
-    "github.com/oskarszura/smarthome/utils"
+    "github.com/oskarszura/smarthome/controllers/utils"
     "github.com/oskarszura/gowebserver/router"
     "github.com/oskarszura/gowebserver/session"
 )
