+++ controllers/dashboard.go
@@ -2,11 +2,11 @@ package controllers
 
 import (
 	"net/http"
-	"github.com/oskarszura/smarthome/utils"
+    "github.com/oskarszura/smarthome/controllers/utils"
 	"github.com/oskarszura/gowebserver/router"
 	"github.com/oskarszura/gowebserver/session"
 )
 
 func CtrDashboard(w http.ResponseWriter, r *http.Request, opt router.UrlOptions, sm session.ISessionManager) {
-	utils.RenderTemplate(w, r, "dashboard", sm)
+    utils.RenderTemplate(w, r, "dashboard", sm)
 }
