+++ models/page.go
@@ -4,4 +4,5 @@ type Page struct {
     Version     string
     Title       string
     IsLogged    bool
+    Params      map[string]interface{}
 }
